Rename source files that don't match the name/nameProvider.go standard.
