+++ providers/_all/all.go
@@ -9,7 +9,7 @@ import (
 	_ "github.com/StackExchange/dnscontrol/providers/digitalocean"
 	_ "github.com/StackExchange/dnscontrol/providers/dnsimple"
 	_ "github.com/StackExchange/dnscontrol/providers/gandi"
-	_ "github.com/StackExchange/dnscontrol/providers/google"
+	_ "github.com/StackExchange/dnscontrol/providers/gcloud"
 	_ "github.com/StackExchange/dnscontrol/providers/namecheap"
 	_ "github.com/StackExchange/dnscontrol/providers/namedotcom"
 	_ "github.com/StackExchange/dnscontrol/providers/ns1"
